// pslpwcl/src/app/modules/pslpapi/model/cpi.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Provincia } from './provincia';

export interface Cpi { 
    codCpi?: string;
    codFp?: number;
    codIntermediarioMin?: string;
    codOperatoreSa?: number;
    codUserAggiorn?: string;
    dsOraGiorni?: string;
    dsSilTCpi?: string;
    flgAttivo?: string;
    flgEnte?: string;
    gruppoOperatoreSa?: string;
    idSilTAgenziaLavoro?: number;
    idSilTCpi?: number;
    idSilTElencoEnti?: string;
    silTProvincia?: Provincia;
    timestamp?: number;
}